separated out Participants Table from Partcipant sections
